Move Segment into SegmentBar, its sole use site
Segment was a single-property-holder type living alone in Chart/Model, one file away from SegmentBar, the only view that renders it. Folding it into SegmentBar.swift keeps the type next to its use site instead of a neutral shared folder. Pure move, no code changes.
